# util/syncutil: don't unconditionally allocate in IntMap.Range