The method was letting its `read` variable escape to the heap, which meant
that the allocation that should have only occurred on the uncommon path where
the dirty map is promoted was being hit on all paths. The effect of this was
that even calls to IntMap.Range on stable maps would allocate.

```
name                       old time/op    new time/op    delta
Range/*syncutil.IntMap-16    1.70µs ±11%    1.65µs ± 9%    -3.15%  (p=0.247 n=10+10)

name                       old alloc/op   new alloc/op   delta
Range/*syncutil.IntMap-16     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
Range/*syncutil.IntMap-16      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```

It looks like this is just an issue with our fork of sync.Map, and it was
introduced when we moved from the use of a `atomic.Value` to an `unsafe.Pointer`
to store the Map's atomically-accessed readOnly struct. Before, the allocation
would be scoped to just the paths that needed to wrap the `readOnly` in an
`interface{}`.

I noticed that the map was the third most frequent source of memory allocation
in a sysbench `oltp_update_non_index` run on a 3x cluster of `r5d.16xlarge` VMs:
```
----------------------------------------------------------+-------------
                                          38371913   100% |   github.com/cockroachdb/cockroach/pkg/storage.(*RaftTransport).queuedMessageCount /go/src/github.com/cockroachdb/cockroach/pkg/storage/raft_transport.go:276
  38371913  1.24%  7.50%   38371913  1.24%                | github.com/cockroachdb/cockroach/pkg/util/syncutil.(*IntMap).Range /go/src/github.com/cockroachdb/cockroach/pkg/util/syncutil/int_map.go:326
----------------------------------------------------------+-------------
```

Release justification: This is about as safe as performance optimizations come.
Others may disagree though, so we can sit on this.

Release note: None