@nvanbenschoten 2019-09-24
> Did this actually move the needle on any high-level (roachtest) perf numbers?

I didn't measure, but I'm pretty sure it will be in the noise on anything at that level. Still, a 1.24% reduction in memory allocations, especially on the hot path in the Raft ready loop isn't negligible. 

This is such a small and safe change that I'm tempted to merge it with the release justification that we'd feel comfortable backporting it. Any objections?