@tbg 2019-09-24
No, after our discussion on the weekly today I also think we should merge.